backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alpaca-trade-api==3.2.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
deprecation==2.1.0
fastapi==0.135.1
frozenlist==1.8.0
h11==0.16.0
idna==3.11
msgpack==1.0.3
multidict==6.7.1
numpy==2.4.2
packaging==26.0
pandas==3.0.1
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.41.0
websocket-client==1.9.0
websockets==10.4
yarl==1.23.0
